suppress warnings when verbose==0
Often, events are saved one-by-one during acquisition. This produces
lots of warnings because all the events before the one current one
are empty in the logFile. One would expect that these warnings can be
suppressed by setting cfg.verbose to 0. Simply adding this condition to
the saveEventsFile does the job.